+++ src/scripts/veaf/veafAirWaves.lua
@@ -1009,7 +1009,7 @@ function AirWaveZone:deployWaves()
         end
         veaf.loggers.get(veafAirWaves.Id):debug("running command [%s]", veaf.lp(command))
         local position = { x = zoneCenter.x - lonDelta, y = zoneCenter.y, z = zoneCenter.z + latDelta }
-        local randomPosition = mist.getRandPointInCircle(position, self.respawnRadius)
+        local randomPosition = veaf.getRandomPointInCircle(position, self.respawnRadius)
         local spawnedGroupsNames = {}
         veafInterpreter.execute(command, randomPosition, self.coalition, nil, spawnedGroupsNames)
         for _, newGroupName in pairs(spawnedGroupsNames) do
@@ -1039,7 +1039,7 @@ function AirWaveZone:deployWaves()
           end
           veaf.loggers.get(veafAirWaves.Id):trace("spawnSpot=%s", veaf.lp(spawnSpot))
           local vars = {}
-          vars.point = mist.getRandPointInCircle(spawnSpot, self.respawnRadius)
+          vars.point = veaf.getRandomPointInCircle(spawnSpot, self.respawnRadius)
           vars.point.z = vars.point.y
           vars.point.y = spawnSpot.y
           vars.gpName = groupName
@@ -1228,7 +1228,7 @@ function AirWaveZone:start()
   -- draw the zone
   if self.drawZone then
     if self.triggerZoneName then
-      self.zoneDrawing = mist.marker.drawZone(self.triggerZoneName, { message = self:getDescription(), readOnly = true })
+      self.zoneDrawing = veaf.drawTriggerZone(self.triggerZoneName, { message = self:getDescription() })
     else
       self.zoneDrawing = VeafCircleOnMap:new()
         :setName(self:getName())
@@ -1254,7 +1254,7 @@ function AirWaveZone:stop()
   -- erase the zone
   if self.zoneDrawing then
     if self.triggerZoneName then
-      mist.marker.remove(self.zoneDrawing.markId)
+      veaf.removeDrawing(self.zoneDrawing.markId)
     else
       self.zoneDrawing:erase()
     end